Android: Fix gamepad L2/R2 digital button bindings.
L2 was previously bound to MISC1 and R2 bound to PADDLE1. This PR
instead brings Android gamepad support (for gamepads that lack
analog triggers) in line with iOS. When a digital L2/R2 is pressed
the buttons trigger a LTRIGGER/RTRIGGER axis event of 1.0, and when
released trigger an event with 0.0. On iOS, this functionality is
built into Apple's GameController framework, on Android we instead
implement this behavior in Godot itself.